requirements.txt
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.2
django-allauth==0.35.0
django-compat==1.0.15
django-hijack==2.1.6
django-hijack-admin==2.1.6
django-import-export==0.7.0
django-markdown-deux==1.0.5
django-silk==2.0.0
django-smtp-ssl==1.0
django-summernote==0.8.8.5
django-taggit==0.22.2
django-taggit-helpers==0.1.4
django-taggit-labels==0.5.1
et-xmlfile==1.0.1
gprof2dot==2017.9.19
idna==2.6
jdcal==1.3
Jinja2==2.10
markdown2==2.3.5
MarkupSafe==1.0
oauthlib==2.0.6
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.0
Pillow==5.0.0
pycodestyle==2.3.1
Pygments==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
